Refactor screen capturing so more is shared between backends
And also fixes the non-working capturing when multiSamples>1

GearVRf-DCO-1.0-Signed-off-by: Mihail Marinov <m.marinov@samsung.com>
